Crate augment_scenario

A tool to modify each person’s schedule from an existing scenario in various ways.

--add_return_trips: For people with only a single trip, add a return trip back home sometime 4-12 hours later. --add_lunch_trips: Before a person’s final trip back home, insert a round-trip to a nearby cafe or restaurant.

These tools aren’t very smart about detecting if a scenario already has these extra trips added in; be careful about running this on the correct input. It modifies the given --input binary scenario in-place.

Functions

add_lunch_trips

add_return_trips

main

pick_lunch_spot

rand_duration